refmt and move helper back to azure_data_cosmos
